Colocate Photos database tests with their modules
