+++ rust/crates/photos/src/db/database.rs
@@ -112,3 +112,74 @@ impl Database {
         })
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::Database;
+    use crate::db::Error;
+
+    const CREATE_ITEMS: &str = "CREATE TABLE items (id INTEGER PRIMARY KEY)";
+
+    #[test]
+    fn reader_queries_reject_writes() {
+        let directory = tempfile::tempdir().unwrap();
+        let db = Database::open(directory.path().join("items.db"), &[CREATE_ITEMS]).unwrap();
+        db.execute("INSERT INTO items (id) VALUES (?)", [7])
+            .unwrap();
+        let deleted: Result<Vec<i64>, _> = db.read_column("DELETE FROM items RETURNING id", ());
+        assert!(matches!(deleted, Err(Error::Sqlite(_))));
+        assert_eq!(
+            db.read_column::<Vec<i64>, _, _>("SELECT id FROM items", ())
+                .unwrap(),
+            [7]
+        );
+    }
+
+    #[test]
+    fn failed_batch_rolls_back_and_writer_recovers() {
+        let directory = tempfile::tempdir().unwrap();
+        let db = Database::open(directory.path().join("items.db"), &[CREATE_ITEMS]).unwrap();
+        assert!(
+            db.write_batch("INSERT INTO items (id) VALUES (?)", [[1], [2], [1]])
+                .is_err()
+        );
+        assert_eq!(
+            db.read_value::<i64>("SELECT COUNT(*) FROM items", ())
+                .unwrap(),
+            0
+        );
+        db.execute("INSERT INTO items (id) VALUES (?)", [3])
+            .unwrap();
+        assert_eq!(
+            db.read_column::<Vec<i64>, _, _>("SELECT id FROM items", ())
+                .unwrap(),
+            [3]
+        );
+    }
+
+    #[test]
+    fn chunked_delete_handles_large_and_empty_id_lists() {
+        let directory = tempfile::tempdir().unwrap();
+        let db = Database::open(directory.path().join("items.db"), &[CREATE_ITEMS]).unwrap();
+        let ids: Vec<i64> = (0..10_001).collect();
+        db.write_batch(
+            "INSERT INTO items (id) VALUES (?)",
+            ids.iter().map(|id| [id]),
+        )
+        .unwrap();
+        db.execute_chunked_in::<i64>("DELETE FROM items WHERE id IN ({})", &[])
+            .unwrap();
+        assert_eq!(
+            db.read_value::<i64>("SELECT COUNT(*) FROM items", ())
+                .unwrap(),
+            10_001
+        );
+        db.execute_chunked_in("DELETE FROM items WHERE id IN ({})", &ids)
+            .unwrap();
+        assert_eq!(
+            db.read_value::<i64>("SELECT COUNT(*) FROM items", ())
+                .unwrap(),
+            0
+        );
+    }
+}
