+++ rust/crates/photos/src/ml_db/caches.rs
@@ -82,3 +82,64 @@ fn now_millis() -> i64 {
         .map(|elapsed| elapsed.as_millis() as i64)
         .unwrap_or(0)
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::db::Connection;
+    use crate::ml_db::tests::{cases, check_seeded, seeded};
+
+    #[test]
+    fn seeded_optional_values() {
+        check_seeded(&cases![
+            "cached face p1": Some("1_0".to_string()) =>
+                |db| db.get_face_id_used_for_person_or_cluster("p1"),
+            "cached face p2": None => |db| db.get_face_id_used_for_person_or_cluster("p2"),
+        ]);
+        check_seeded(&cases![
+            "cached dog": Some(vec![0.5, -1.0]) => |db| db.get_repeated_text_embedding_cache("dog"),
+            "uncached cat": None => |db| db.get_repeated_text_embedding_cache("cat"),
+        ]);
+    }
+
+    #[test]
+    fn text_embedding_cache_replaces_and_expires() {
+        let (directory, db) = seeded();
+        db.put_repeated_text_embedding_cache("dog", &[2.0]).unwrap();
+        assert_eq!(
+            db.get_repeated_text_embedding_cache("dog").unwrap(),
+            Some(vec![2.0f32])
+        );
+
+        let connection = Connection::open(directory.path().join("ente.ml.db")).unwrap();
+        connection
+            .execute(
+                "UPDATE text_embeddings_cache SET created_at = 1 WHERE text_query = ?",
+                ["dog"],
+            )
+            .unwrap();
+        assert_eq!(db.get_repeated_text_embedding_cache("dog").unwrap(), None);
+        let remaining: i64 = connection
+            .query_row("SELECT COUNT(*) FROM text_embeddings_cache", (), |row| {
+                row.get(0)
+            })
+            .unwrap();
+        assert_eq!(remaining, 0);
+    }
+
+    #[test]
+    fn face_cache_replaces_and_removes() {
+        let (_directory, db) = seeded();
+        db.put_face_id_cached_for_person_or_cluster("p1", "2_0")
+            .unwrap();
+        assert_eq!(
+            db.get_face_id_used_for_person_or_cluster("p1").unwrap(),
+            Some("2_0".to_string())
+        );
+        db.remove_face_id_cached_for_person_or_cluster("p1")
+            .unwrap();
+        assert_eq!(
+            db.get_face_id_used_for_person_or_cluster("p1").unwrap(),
+            None
+        );
+    }
+}
