+++ rust/crates/photos/src/ml_db/clip.rs
@@ -105,3 +105,88 @@ fn clip_row(embedding: &ClipEmbedding) -> (i64, Vec<u8>, i64) {
         embedding.version,
     )
 }
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use super::ClipRow;
+    use crate::ml_db::codec::encode_f32;
+    use crate::ml_db::constants::CLIP_EMBEDDING_DIMENSIONS;
+    use crate::ml_db::tests::{cases, check_seeded, clip, full_clip, seeded};
+
+    #[test]
+    fn seeded_counts() {
+        check_seeded(&cases![
+            "clip files": 4 => |db| db.get_clip_indexed_file_count(1),
+            "clip files v2": 1 => |db| db.get_clip_indexed_file_count(2),
+            "clip vectorizable": 2 => |db| db.get_clip_vectorizable_file_count(1),
+            "clip vectorizable v2": 1 => |db| db.get_clip_vectorizable_file_count(2),
+            "clip rows": 4 => |db| db.count_clip_rows(),
+        ]);
+    }
+
+    #[test]
+    fn seeded_versions() {
+        check_seeded(&cases![
+            "clip versions": HashMap::from([(1, 1), (2, 2), (3, 1), (4, 1)]) =>
+                |db| db.clip_indexed_file_with_version(),
+        ]);
+    }
+
+    #[test]
+    fn seeded_rows() {
+        let (_directory, db) = seeded();
+        let mut vectors = db.get_all_clip_vectors().unwrap();
+        vectors.sort_by_key(|vector| vector.file_id);
+        let dimensions: Vec<(i64, usize)> = vectors
+            .iter()
+            .map(|v| (v.file_id, v.embedding.len()))
+            .collect();
+        assert_eq!(
+            dimensions,
+            vec![
+                (1, CLIP_EMBEDDING_DIMENSIONS),
+                (2, CLIP_EMBEDDING_DIMENSIONS),
+                (3, 2)
+            ]
+        );
+        assert_eq!(vectors[2].embedding, vec![1.0f32, 2.0]);
+        let page = db.get_clip_rows_page(2, 1).unwrap();
+        let clip_row = |file_id, embedding: Vec<f32>| ClipRow {
+            file_id,
+            embedding: encode_f32(embedding),
+        };
+        assert_eq!(
+            page,
+            vec![
+                clip_row(3, vec![1.0, 2.0]),
+                clip_row(2, vec![0.25; CLIP_EMBEDDING_DIMENSIONS])
+            ]
+        );
+    }
+
+    #[test]
+    fn clip_rows_replace_and_delete() {
+        let (_directory, db) = seeded();
+        db.insert_clip_rows(&[]).unwrap();
+        db.insert_clip_rows(&[clip(1, vec![1.0])]).unwrap();
+        db.insert_clip_rows(&[clip(3, vec![3.0]), full_clip(5)])
+            .unwrap();
+        assert_eq!(
+            db.clip_indexed_file_with_version().unwrap(),
+            HashMap::from([(1, 1), (2, 2), (3, 1), (4, 1), (5, 1)])
+        );
+        let mut vectors = db.get_all_clip_vectors().unwrap();
+        vectors.sort_by_key(|vector| vector.file_id);
+        assert_eq!(vectors[0].embedding, vec![1.0f32]);
+        assert_eq!(vectors[2].embedding, vec![3.0f32]);
+        assert_eq!(db.get_clip_vectorizable_file_count(1).unwrap(), 2);
+
+        db.delete_clip_rows(&[]).unwrap();
+        db.delete_clip_rows(&[1, 3]).unwrap();
+        assert_eq!(db.count_clip_rows().unwrap(), 3);
+        db.delete_all_clip_rows().unwrap();
+        assert_eq!(db.count_clip_rows().unwrap(), 0);
+    }
+}
